// src/app/model/theme/ThemeColorSet.d.ts
export type ThemeColorSet = {
  primaryBackgroundColor: string;
  primaryForegroundColor: string;
  primaryHovergroundColor: string;
  primaryActivegroundColor: string;
  primaryActivegroundFontColor: string;
  primaryBordergroundColor: string;
  primaryBoxShadow: string;
  primaryChatShadow: string;
  primarySettingShadow: string;
};